third_party/blink/web_tests/platform/generic/http/tests/websocket/bad-sub-protocol-control-chars-expected.txt

Test WebSocket bad sub-protocol names by control characters.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new WebSocket("ws://127.0.0.1:8880/simple", "\0") threw exception SyntaxError: Failed to construct 'WebSocket': The subprotocol '\u0000' is invalid..
PASS new WebSocket("ws://127.0.0.1:8880/simple", "	") threw exception SyntaxError: Failed to construct 'WebSocket': The subprotocol '\u0009' is invalid..
PASS new WebSocket("ws://127.0.0.1:8880/simple", "
") threw exception SyntaxError: Invalid or unexpected token.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "
") threw exception SyntaxError: Invalid or unexpected token.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "") threw exception SyntaxError: Failed to construct 'WebSocket': The subprotocol '\u001B' is invalid..
PASS successfullyParsed is true

TEST COMPLETE
